extract runEquation, handleDoubleEquals, handleNestedParens
